refactor(menus): domain 패키지 내에서 exception, model, repository 로 세부 분류
