+++ src/main/java/kitchenpos/menus/tobe/domain/exception/ProfaneNameException.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.exception;
 
 public class ProfaneNameException extends IllegalArgumentException {
 }
