+++ src/main/java/kitchenpos/menus/tobe/domain/model/Profanity.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.model;
 
 @FunctionalInterface
 public interface Profanity {
